Adjust behaviour for SSL_ERROR_SYSCALL
As per https://www.openssl.org/docs/man1.1.1/man3/SSL_get_error.html

> The SSL_ERROR_SYSCALL with errno value of 0 indicates unexpected EOF
> from the peer. This will be properly reported as SSL_ERROR_SSL with
> reason code SSL_R_UNEXPECTED_EOF_WHILE_READING in the OpenSSL 3.0
> release because it is truly a TLS protocol error to terminate the
> connection without a SSL_shutdown().
>
> The issue is kept unfixed in OpenSSL 1.1.1 releases because many
> applications which choose to ignore this protocol error depend on the
> existing way of reporting the error.

I think trilogy would be correct in turning this into an OpenSSL error
(as will happen on openssl 3.0) but for improved error reporting and
restoring closer to previous behaviour we can treat this as a connection
closed.

Additionally, this now only translates an exception to SSL_ERROR_SYSCALL
if there is nothing in the error queue.