Shutdown connection on external exception
When we call rb_wait_for_single_fd we release the GVL and allow our
Ruby thread to receive interrupts. The most obvious case of this is
using Timeout.timeout.

If we see an exception in this way we can't safely use our socket
anymore as it was likely in the middle of another operation, so it
should be shut down.